+++ examples/src/example-dynamic-navmesh.ts
@@ -1,6 +1,6 @@
 import Rapier from '@dimforge/rapier3d-compat';
 import { GUI } from 'lil-gui';
-import { box3, triangle3, type Vec3, vec2, vec3 } from 'mathcat';
+import { box3, type Vec3, vec2, vec3 } from 'mathcat';
 import {
     addOffMeshConnection,
     addTile,
@@ -428,7 +428,9 @@ function initDynamicNavMesh(
     const tileStaticHeightfields = new Map<string, ReturnType<typeof createHeightfield>>();
 
     const borderOffset = config.borderSize * config.cellSize;
-    const triangle = triangle3.create();
+    const triA: Vec3 = [0, 0, 0];
+    const triB: Vec3 = [0, 0, 0];
+    const triC: Vec3 = [0, 0, 0];
 
     for (let tx = 0; tx < tileWidth; tx++) {
         for (let ty = 0; ty < tileHeight; ty++) {
@@ -455,11 +457,11 @@ function initDynamicNavMesh(
                 const b = levelIndices[i + 1];
                 const c = levelIndices[i + 2];
 
-                vec3.fromBuffer(triangle[0], levelPositions, a * 3);
-                vec3.fromBuffer(triangle[1], levelPositions, b * 3);
-                vec3.fromBuffer(triangle[2], levelPositions, c * 3);
+                vec3.fromBuffer(triA, levelPositions, a * 3);
+                vec3.fromBuffer(triB, levelPositions, b * 3);
+                vec3.fromBuffer(triC, levelPositions, c * 3);
 
-                if (box3.intersectsTriangle3(expandedBox, triangle)) {
+                if (box3.intersectsTriangle3(expandedBox, triA, triB, triC)) {
                     trianglesInBox.push(a, b, c);
                 }
             }